Fix missing signing and upload properties issue
When developer does not have signing and/or sonatype properties set in
their environment the gradle script used to blow up.
I have 'fixed' this by providing default values for sonatypeUsername
and sonatypePassword if they do not exist. In regards to signing
the workaround is to only require signing when doing a release.