Set project name rather than archivesBaseName
Project name is a read only property in build.gradle, but it can be set
in settings.gradle. This makes things more uniform and deterministic.
Project now has the same name no matter which directory you check it
out to.
It also helps with Intellij IDEA not creating additional modules and
subsequently having conflicts between them.