Add LoadSensitiveWait Strategy.
LoadSensitive strategy is the combined strategy that use BusySpin and
BlockingWait.
It achieves similar high throughput and low latency as BusySpin Strategy
in high load scenarios, while performs like BlockingWait strategy in
idle state to save CPU power and system resource.
The basic concept is a bit similar with PhasedBackOffWait Strategy, but
the performance is better and easier to use without customizations.