Try a different CI action
The previous CI action published the whole git project to the branch,
only replacing the `.adoc` files with `.html` versions. This meant that
the actual ASCIIDoc site lived in a weird subdirectory, which isn't
ideal. This action seems to suggest that it will only publish a certain
directory, which might be better for what we want?