`Sequence::addAndGet` should return the sequence value after the addition.
This was broken since f6d4bf94a802926de9d6ed7f0f7ac9de254a167e and a unit test has been added to make sure the same mistake doesn't happen again.
The `VarHandle::getAndAdd` "returns the variable's previous value, as accessed with the memory semantics of getVolatile", so there is no need for a manual version with a full barrier.
This change improves the performance of `Sequence::addAndGet` & `Sequence::incrementAndGet` as seen in the results of SequenceBenchmark before/after this change.