Extract examples from documentation
Previously, the example code was all written directly into the
documentation, decoupling it from any real code. This meant changes to
APIs or whatever could unknowningly break the example code, adding to
the number of considerations that a developer would, in theory, be
required to worry about when changing something.

By promoting these examples to real code, we can offload that mental
strain onto the compiler, thus giving us more confidence that the
examples work. In theory, it also means that in the future we could
even add some tests for these to prove they're doing what we claim!