Remove use of unsafe from RingBuffer
Unsafe was used to get a speed boost by having direct array access, skipping bounds checks.
While use of Unsafe is now deprecated I don't think there is a modern equivalent to the old array access methods.
Downside: Going back to a plain array and array access is slower
Upsides: Not relying on Unsafe; The code is a lot simpler