Assert that only a single thread accesses a SingleProducerSequencer.
If a developer uses the wrong ProducerType, it's hard to detect.
This commit adds a check using assertions so a developer can fail in
test instead of being burnt in Production.
(There will be no runtime cost in Production - only an unused field.)
I believe this safety check is worth that cost.
As can be seen from the commit, there was actually a test case that
started failing because a SingleProducerSequencer was accessed by two
threads...